requirements.txt
streamlit
langchain_openai
langchain.prompts
langchain.schema